requirements.txt
numpy>=1.23
scipy>=1.9
PySide6>=6.6
matplotlib>=3.8
scikit-image>=0.22
Pillow>=10.1
